Tweaks and optimization
